refactor: Diet 도메인 재구조화, DTO 패키지 구조 리팩토링
