+++ Backend/src/main/java/com/demo/nimn/controller/diet/DailyDietController.java
@@ -1,7 +1,7 @@
 package com.demo.nimn.controller.diet;
 
-import com.demo.nimn.dto.diet.Response.DailyDietDTO;
-import com.demo.nimn.dto.diet.Request.DailyRequestDTO;
+import com.demo.nimn.dto.diet.DailyDietDTO;
+import com.demo.nimn.dto.diet.DailyRequestDTO;
 import com.demo.nimn.service.diet.DietService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
