+++ Backend/src/main/java/com/demo/nimn/dto/diet/DailyDietDTO.java
@@ -1,4 +1,4 @@
-package com.demo.nimn.dto.diet.Response;
+package com.demo.nimn.dto.diet;
 
 import lombok.*;
