+++ Backend/src/main/java/com/demo/nimn/dto/diet/DailyDietRequestDTO.java
@@ -1,4 +1,4 @@
-package com.demo.nimn.dto.diet.Request;
+package com.demo.nimn.dto.diet;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
