+++ Backend/src/main/java/com/demo/nimn/dto/diet/DailyRequestDTO.java
@@ -1,4 +1,4 @@
-package com.demo.nimn.dto.diet.Request;
+package com.demo.nimn.dto.diet;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,13 +6,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.util.List;
 
-@Data
 @Builder
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserDailyMealPlanDTO {
+public class DailyRequestDTO {
+    String userEmail;
     LocalDate date;
-    List<Long> foodMenuIds;
 }
