+++ Backend/src/main/java/com/demo/nimn/dto/diet/FoodSelectionRequestDTO.java
@@ -1,4 +1,4 @@
-package com.demo.nimn.dto.diet.Request;
+package com.demo.nimn.dto.diet;
 
 import lombok.*;
