+++ Backend/src/main/java/com/demo/nimn/service/diet/DietService.java
@@ -1,9 +1,9 @@
 package com.demo.nimn.service.diet;
 
-import com.demo.nimn.dto.diet.Request.WeeklyDietRequestDTO;
-import com.demo.nimn.dto.diet.Response.DailyDietDTO;
-import com.demo.nimn.dto.diet.Request.DailyDietRequestDTO;
-import com.demo.nimn.dto.diet.Response.WeeklyDietDTO;
+import com.demo.nimn.dto.diet.WeeklyDietRequestDTO;
+import com.demo.nimn.dto.diet.DailyDietDTO;
+import com.demo.nimn.dto.diet.DailyDietRequestDTO;
+import com.demo.nimn.dto.diet.WeeklyDietDTO;
 import com.demo.nimn.entity.diet.DailyDiet;
 
 import java.time.LocalDate;
