+++ Backend/src/main/java/com/demo/nimn/service/diet/FoodSelectionServiceImpl.java
@@ -1,8 +1,7 @@
 package com.demo.nimn.service.diet;
 
-import com.demo.nimn.dto.diet.Response.FoodSelectionDTO;
-import com.demo.nimn.dto.diet.Request.FoodSelectionRequestDTO;
-import com.demo.nimn.dto.food.FoodDTO;
+import com.demo.nimn.dto.diet.FoodSelectionDTO;
+import com.demo.nimn.dto.diet.FoodSelectionRequestDTO;
 import com.demo.nimn.entity.diet.FoodSelection;
 import com.demo.nimn.entity.food.Food;
 import com.demo.nimn.repository.diet.FoodSelectionRepository;
