+++ Backend/src/main/java/com/demo/nimn/service/review/ReviewService.java
@@ -1,6 +1,6 @@
 package com.demo.nimn.service.review;
 
-import com.demo.nimn.dto.diet.Response.WeeklyDietDTO;
+import com.demo.nimn.dto.diet.WeeklyDietDTO;
 import com.demo.nimn.dto.review.DailyDietReviewDTO;
 import com.demo.nimn.dto.review.ReviewDTO;
 import com.demo.nimn.dto.review.ReviewSummaryDTO;
